+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Updated the `bitmovin-player-react-native` dependency to `0.27.1`.
+
 ## [1.0.0] - 2024-06-14
 
 Adds support for Conviva Analytics with Bitmovin Player React Native SDK.