+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Updated the `bitmovin-player-react-native` dependency to `0.27.1`.
 - Updated the Bitmovin Player Conviva Analytics Integration for Android dependency to `2.5.0`.
+- Updated the Bitmovin Player Conviva Analytics Integration for iOS dependency to `3.3.0`.
 
 ## [1.0.0] - 2024-06-14